backend/requirements.txt
anyio==4.11.0
click==8.3.1
colorama==0.4.6
fastapi==0.95.2
h11==0.16.0
httptools==0.7.1
idna==3.11
psycopg==3.2.12
psycopg-binary==3.2.12
psycopg2-binary==2.9.11
pydantic==1.10.24
python-dotenv==1.2.1
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.21.1
watchfiles==1.1.1
websockets==15.0.1
